+++ CHANGELOG.md
@@ -12,6 +12,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #726 Added @Preauthorize annotation to dashboard controller
 ### Changed
 - #844 Prefilled bpn on investigation creation
+- #843 Refactored e2e tests, added edit notification e2e test case
 - #828 fix duplicates in traction_battery_code_subcomponent table
 ### Removed
 -